requirements.txt
moviepy
mutagen
requests
bs4
ollama
openai
ushlex
openai-whisper